requirements.txt
flask
scikit-learn
joblib
fpdf
requests
python-dotenv
firebase-admin
firebase
pandas
numpy
